refactor: move metadata functions to a unified module
