+++ crypto-ffi/src/lib.rs
@@ -22,8 +22,6 @@ cfg_if::cfg_if! {
         mod wasm;
         pub use self::wasm::*;
     } else {
-        uniffi::setup_scaffolding!("core_crypto_ffi");
-
         mod generic;
         pub use self::generic::*;
     }
@@ -31,3 +29,10 @@ cfg_if::cfg_if! {
 
 #[cfg(doc)]
 pub mod bindings;
+
+#[cfg(not(target_family = "wasm"))]
+uniffi::setup_scaffolding!("core_crypto_ffi");
+
+mod metadata;
+
+pub use metadata::{BuildMetadata, build_metadata, version};
